Initial commit of my work on the Udemy web development course with Colt Steele.
I am starting off with a project that is later in the course called,
YelpCamp. This is my first cut of code (it is not nearly done at all).

I just want to see if I can upload this repo to GitHub then clone it
to my linux laptop and vice versa.