Completely re-worked the campground show (details) page to put a single image of the campground in a Bootstrap grid with a nice caption box with price/night, and the name of the campground. Also, there is a bootstrap well I added that contains comment by user.
Along the left side of the campground details page is a list of items in
a bootstrap grid. These are placeholders for use later.

Added styling for thumbnail images (on index page and campground show (details)
page) so that they are as wide as the div they reside in plus have some nice
padding around them so that text below is not too close.

To support this styling I had to add classes to some div or img html
elements.