(express.js) insert middleware for exporting currentUser name. we'll need to access this in any rendered view. For now, we're just using in in the partials header to determine if a user is logged in or not.
(header.ejs) - if there is no user signed in then show options in navbar
to login or register. if a user IS logged in then only show the logout
option in the navbar.

(comments.ejs) - check if user is logged in before allowing a post of a new
comment. This is additional checking just to be sure. we also (previous commit)
disallow users to even get to the new comment page if they are not logged in.

(index.js) - when logging out only execute req.logout() if the current user
name is defined.